move `get_stats_record` helper fn from schema.rs to util.rs so qsvdp can compile
and we don't have to include the `schema` cmd into qsvdp
